requirements.txt
pymongo==4.10.1
python-dotenv==1.0.1
Flask==3.0.3
bcrypt==4.2.0
gunicorn==23.0.0